Fix/refactor call handling (#6)
* feat: :adhesive_bandage: Add-arrayLocalICECandidates

* refactor: :adhesive_bandage: startCamera-in-utils-peerConnection inside component
